refactor: move `Catalog` to pages layer
